requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.0
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
